Got the traceback error to go away
You were close! The "w" is what tells a file to open for writing instead of reading.

If you run this you'll see that the output comes out one character per line, which is probably not what you intended. This is because the file.read() produces a string, so you're actually working character by character not line by line.

Try using "for data in file:" instead of "data=file.read(...) ... for line in data:". You'll still have to do the replace() call inside the loop since you're going line by line instead of first doing it over the entire string.